refactor(#27): split vessel drafter model into focused sub-modules
Partially addresses issue #27 by splitting the 474-LOC
``models/vessel_drafter.py`` module along its natural seams:

- ``models/vessel_drafter_types.py`` (120 LOC) holds the frozen
  dataclasses for material layers, radial bands, electrode placements,
  and side/lid ports.
- ``models/vessel_drafter_manifest.py`` (146 LOC) holds the private
  ``_build_*_manifest`` helpers used by ``VesselDrafterLayout.to_manifest``.
- ``models/vessel_drafter.py`` now hosts only ``VesselDrafterLayout``
  and re-exports every public name plus the manifest helpers to keep the
  import surface stable for callers and existing tests.

Added ``tests/test_vessel_drafter_types.py`` with direct coverage of the
extracted dataclasses. The existing manifest tests continue to import
the ``_build_*`` helpers from ``vessel_drafter`` unchanged.
